+++ app/src/state/non-action-util.ts
@@ -147,3 +147,25 @@ export function getPastWorkshops(workshops: Workshop[]) {
 export function getMapImageUrl(mapInfo: MapInfo[string]) {
   return devSafeUrl(`/map-image/${mapInfo.datapackTitle}/${mapInfo.uuid}/${mapInfo.img}`);
 }
+export async function downloadFile(blob: Blob, filename: string) {
+  const reader = new FileReader();
+  reader.readAsDataURL(blob);
+  await new Promise((resolve, reject) => {
+    reader.onloadend = resolve;
+    reader.onerror = reject;
+  });
+  if (typeof reader.result !== "string") {
+    throw new Error("Invalid file");
+  }
+  const fileURL = reader.result;
+  if (!fileURL) throw new Error("Invalid file");
+  const aTag = document.createElement("a");
+  aTag.href = fileURL;
+
+  aTag.setAttribute("download", filename);
+
+  document.body.appendChild(aTag);
+  aTag.click();
+  aTag.remove();
+  document.removeChild(aTag);
+}
