+++ app/src/util/error-codes.ts
@@ -96,7 +96,9 @@ export enum ErrorCodes {
   ADMIN_WORKSHOP_ENDED = "ADMIN_WORKSHOP_ENDED",
   WORKSHOP_FETCH_DATAPACK_FAILED = "WORKSHOP_FETCH_DATAPACK_FAILED",
   METADATA_NOT_FOUND = "METADATA_NOT_FOUND",
-  INVALID_CROSSPLOT_CONVERSION = "INVALID_CROSSPLOT_CONVERSION"
+  INVALID_CROSSPLOT_CONVERSION = "INVALID_CROSSPLOT_CONVERSION",
+  INVALID_CROSSPLOT_UNITS = "INVALID_CROSSPLOT_UNITS",
+  CROSSPLOT_CONVERSION_FAILED = "CROSSPLOT_CONVERSION_FAILED"
 }
 
 export const ErrorMessages = {
@@ -197,5 +199,7 @@ export const ErrorMessages = {
   [ErrorCodes.ADMIN_WORKSHOP_ENDED]: i18n.t("errors.ADMIN_WORKSHOP_ENDED"),
   [ErrorCodes.WORKSHOP_FETCH_DATAPACK_FAILED]: i18n.t("errors.WORKSHOP_FETCH_DATAPACK_FAILED"),
   [ErrorCodes.METADATA_NOT_FOUND]: i18n.t("errors.METADATA_NOT_FOUND"),
-  [ErrorCodes.INVALID_CROSSPLOT_CONVERSION]: i18n.t("errors.INVALID_CROSSPLOT_CONVERSION")
+  [ErrorCodes.INVALID_CROSSPLOT_CONVERSION]: i18n.t("errors.INVALID_CROSSPLOT_CONVERSION"),
+  [ErrorCodes.INVALID_CROSSPLOT_UNITS]: i18n.t("errors.INVALID_CROSSPLOT_UNITS"),
+  [ErrorCodes.CROSSPLOT_CONVERSION_FAILED]: i18n.t("errors.CROSSPLOT_CONVERSION_FAILED")
 };
